backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
xgboost==2.0.2
requests==2.31.0
python-multipart==0.0.6
pydantic==2.5.0
python-dateutil==2.8.2
holidays==0.37
sqlalchemy==2.0.23
alembic==1.13.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-decouple==3.8
psycopg2-binary==2.9.9